requirements.txt
anthropic==0.39.0
datasketch==1.6.4
faiss-cpu==1.8.0
feedparser==6.0.11
hdbscan==0.8.38
scipy==1.17.1
networkx==3.3
numpy==1.26.4
pandas==2.2.2
playwright==1.44.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
sec-edgar-downloader==5.0.0
sentence-transformers==3.0.1
websockets==16.0
yfinance==1.2.0
praw==7.8.1
